pyproject.toml
[project]
name = "mrql"
version = "0.1.0"
description = "MongoDB Readable Query Language"
authors = [{ name = "Denis Kazakov", email = "denis@kazakov.ru.net" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lark-parser>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/KazakovDenis/mrql"
Repository = "https://github.com/KazakovDenis/mrql"

[dependency-groups]
dev = [
    "ruff>=0.13.0",
    "ty>=0.0.1a20",
]
